backend/requirements.txt
# Existing
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
python-multipart==0.0.6
python-dotenv==1.0.0
email-validator==2.1.0

# NEW: For scanning features
requests==2.31.0
aiohttp==3.9.1
dnspython==2.4.2
whois==0.9.27
python-whois==0.8.0
shodan==1.29.1
censys==2.2.9
nmap==1.6.2
beautifulsoup4==4.12.2
lxml==4.9.3